requirements.txt
streamlit
fastmcp
ollama
pydantic
pandas
plotly
